refactor: relocate virtualizer to vaadin-virtual-list package
