+++ packages/vaadin-virtual-list/virtualizer.js
@@ -1,0 +1,1 @@
+export * from './src/virtualizer.js';
