--- packages/vaadin-virtualizer/vaadin-virtualizer.js
@@ -1,1 +1,0 @@
-export * from './src/vaadin-virtualizer.js';
